+++ buildpacks/ruby/src/main.rs
@@ -6,7 +6,6 @@
 use bullet_stream::{global::print, style};
 use commons::cache::CacheError;
 use commons::gemfile_lock::GemfileLock;
-use commons::metadata_digest::MetadataDigest;
 use core::str::FromStr;
 use fs_err::PathExt;
 use fun_run::CmdError;
@@ -207,21 +206,6 @@ impl Buildpack for RubyBuildpack {
                     },
                     cpu_architecture: context.target.arch.clone(),
                     ruby_version: ruby_version.clone(),
-                    force_bundle_install_key: String::from(
-                        crate::layers::bundle_install_layer::FORCE_BUNDLE_INSTALL_CACHE_KEY,
-                    ),
-                    digest: MetadataDigest::new_env_files(
-                        &context.platform,
-                        &[
-                            &context.app_dir.join("Gemfile"),
-                            &context.app_dir.join("Gemfile.lock"),
-                        ],
-                    )
-                    .map_err(|error| match error {
-                        commons::metadata_digest::DigestError::CannotReadFile(path, error) => {
-                            RubyBuildpackError::BundleInstallDigestError(path, error)
-                        }
-                    })?,
                 },
                 &BundleWithout::new("development:test"),
             )?;
@@ -291,7 +275,6 @@ pub(crate) enum RubyBuildpackError {
     MetricsAgentError(MetricsAgentInstallError),
     MissingGemfileLock(std::path::PathBuf, std::io::Error),
     InAppDirCacheError(CacheError),
-    BundleInstallDigestError(std::path::PathBuf, std::io::Error),
     BundleInstallCommandError(CmdError),
     RakeAssetsPrecompileFailed(CmdError),
     GemInstallBundlerCommandError(CmdError),